moved chat index into /chat
